+++ web/packages/teleport/src/Bots/Add/AddBots.tsx
@@ -17,11 +17,10 @@
  */
 
 import { FeatureBox } from 'teleport/components/Layout';
-import { Route, Switch } from 'teleport/components/Router';
+import { Redirect, Route, Switch } from 'teleport/components/Router';
 import cfg from 'teleport/config';
 
 import { BotFlowType } from '../types';
-import { AddBotsPicker } from './AddBotsPicker';
 import GitHubActionsFlow from './GitHubActions';
 
 export function AddBots() {
@@ -32,7 +31,12 @@ export function AddBots() {
           path={cfg.getBotsNewRoute(BotFlowType.GitHubActions)}
           component={GitHubActionsFlow}
         />
-        <Route path={cfg.getBotsNewRoute()} component={AddBotsPicker} />
+        <Route
+          path={cfg.getBotsNewRoute()}
+          component={() => (
+            <Redirect to={`${cfg.getIntegrationEnrollRoute()}?kinds=bot`} />
+          )}
+        />
       </Switch>
     </FeatureBox>
   );
