+++ web/packages/teleport/src/config.ts
@@ -825,10 +825,7 @@ const cfg = {
   },
 
   getBotsNewRoute(type?: string) {
-    let route = generatePath(
-      generatePath(cfg.routes.integrationEnroll, { type })
-    );
-    return `${route}?kinds=bot`;
+    return generatePath(cfg.routes.botsNew, { type });
   },
 
   getAppsRoute(clusterId: string) {
