+++ web/packages/teleport/src/features.tsx
@@ -346,7 +346,7 @@ export class FeatureAddBotsShortcut implements TeleportFeature {
     icon: AddCircle,
     exact: true,
     getLink() {
-      return cfg.getBotsNewRoute();
+      return `${cfg.getIntegrationEnrollRoute()}?kinds=bot`;
     },
   };
 }
